serialize run

+++ src/main/python/auto-calibration/scripts/correlational.py
@@ -1,6 +1,5 @@
 # Author: Kiran Chhatre
-# Implementation 2 related
-import os, subprocess, time, glob, csv, shutil, fnmatch
+import os, subprocess, time, glob, csv, shutil
 import pandas as pd
 from modify_csv import modify_csv
 from config import *
@@ -9,9 +8,12 @@ from worker import ext_change, change_conf
 # KEEP ALL INTERCEPTS AS ZERO and KEEP OUTPUT FOLDER EMPTY!!
 
 # Deleting shared o/p folder contents
-filelist = [ f for f in os.listdir(shared) ]
-for f in filelist:
-    os.remove(os.path.join(shared, f))
+if not os.path.exists(shared):
+    os.makedirs(shared)
+else:
+    file_list = [f for f in os.listdir(shared)]
+    for f in file_list:
+        os.remove(os.path.join(shared, f))
 
 ################################### Preprocessing
 
@@ -19,11 +21,9 @@ num = 1
 shutil.copy(base_urbansim_config, copy_urbansim_config % (num))
 picked_conf_file = copy_urbansim_config % (num)   # label the file
 filename = copy_urbansim_txt % (num)
-input_vector = [0,0,0,0,0,0,0,0,0,0]
-#input_vector = [-12,1.5,6.25,0.5,-7.5,0,14.25,4.75] # in case you want to initiate at best intercepts
-finaliteration = '0'
+input_vector = [0, 0, 0, 0, 0, 0, 0, 0, 1]
 ext_change('edit', picked_conf_file, filename)
-change_conf(input_vector=input_vector, filename=filename)
+change_conf(input_vector, filename)
 ext_change('save', picked_conf_file, filename)
 
 ################################### Fire BEAM
@@ -34,23 +34,25 @@ os.chdir(search_space)
 ################################### Bookkeeping phase
 
 out_dir = glob.glob(sf_light_dir)
-
-while not out_dir:
+out_file = out_dir[0]+'/referenceRealizedModeChoice.csv'
+while not os.path.exists(out_file):
     time.sleep(1)
 
-out_file = out_dir[0]+'/referenceRealizedModeChoice.csv'
+df = pd.read_csv(out_file)
 
-
-if os.path.isfile(out_file):
-    df = pd.read_csv(out_file)
-else:
-    raise ValueError("%s isn't a file!" % file_path)
-
-df.loc[1,'iterations'] = 'modeshare_now'
+df.loc[1, 'iterations'] = 'modeshare_now'
 del df['cav']
+
+car_speed = out_dir[0]+'/CarSpeed.csv'
+free_flow_car_speed = out_dir[0]+'/FreeFlowCarSpeed.csv'
+car_speed_df = pd.read_csv(car_speed)
+car_speed_avg=car_speed_df[(car_speed_df["iteration"] == finaliteration) | (car_speed_df["carType"] == 'Personal')]['avg'].values[0]
+car_speed_df = pd.read_csv(free_flow_car_speed)
+free_flow_car_speed_avg=car_speed_df[(car_speed_df["iteration"] == finaliteration_flowcapacity) | (car_speed_df["carType"] == 'Personal')]['avg'].values[0]
+df['speed'] = [free_flow_car_speed_avg, car_speed_avg]
+
 input_vector.insert(0, "intercepts_now")
 df.loc[-1] = input_vector
-#df.loc[-1] = ['intercepts_now', 0,0,0,0,0,0,0,0]
 df.index = df.index+1
 df.sort_index(inplace=True)
 df.set_index('iterations', inplace=True)
@@ -63,3 +65,4 @@ total_L1 = df.loc['L1'].abs().sum()
 df.to_csv(output_csv % (1, total_L1), sep='\t', encoding='utf-8')
 csv_name = output_csv % (1, total_L1)
 modify_csv(csv_name=csv_name)
+
